requirements.txt
Jinja2==3.1.4
ninja==1.11.1.1
numpy==1.26.4
pillow==11.0.0
tensorboard==2.16.2
torch==2.1.0+cu118
torchaudio==2.1.0+cu118
torchvision==0.16.0+cu118
tqdm==4.67.1
accelerate==1.5.2
transformers==4.49.0
datasets==3.4.1
peft==0.14.0
diffusers==0.32.2
sentencepiece==0.2.0
simple-aesthetics-predictor==0.1.2
image-reward==1.5
wandb==0.19.8
trl==0.15.2
scikit-learn==1.6.1
